Separate Flight integration tests into new modules
This refactor is a prerequisite to adding Flight SQL integration tests, as the current integration tests do not allow Flight SQL usage (because it would cause circular dependencies)
