+++ cpp/CMakeLists.txt
@@ -334,6 +334,10 @@ if(ARROW_GANDIVA)
   set(ARROW_WITH_RE2 ON)
 endif()
 
+if(ARROW_BUILD_INTEGRATION AND ARROW_FLIGHT)
+  set(ARROW_FLIGHT_SQL ON)
+endif()
+
 if(ARROW_FLIGHT_SQL)
   set(ARROW_FLIGHT ON)
 endif()
