+++ cpp/src/arrow/CMakeLists.txt
@@ -730,6 +730,9 @@ endif()
 
 if(ARROW_FLIGHT)
   add_subdirectory(flight)
+  if(ARROW_BUILD_INTEGRATION)
+    add_subdirectory(flight/integration_tests)
+  endif()
 endif()
 
 if(ARROW_FLIGHT_SQL)
