+++ cpp/src/arrow/flight/sql/test_server_cli.cc
@@ -20,11 +20,10 @@
 #include <iostream>
 #include <memory>
 #include <string>
+#include <csignal>
 
 #include "arrow/flight/server.h"
 #include "arrow/flight/sql/example/sqlite_server.h"
-#include "arrow/flight/test_integration.h"
-#include "arrow/flight/test_util.h"
 #include "arrow/io/test_common.h"
 #include "arrow/testing/json_integration.h"
 #include "arrow/util/logging.h"
