+++ java/flight/pom.xml
@@ -33,6 +33,7 @@
         <module>flight-core</module>
         <module>flight-grpc</module>
         <module>flight-sql</module>
+        <module>flight-integration-tests</module>
     </modules>
 
     <build>
